feat(welcome): welcome screen using component lib
This patch turns this nextjs application into an importable
lib of (untranspiled) components.

The first exported component is a react component with the
contents of the welcome screen.

This patch also replaces the default Nextjs boilerplate home
page with a welcome screen using this component.

Finally, it updates the README to state that the components can be used by other nextjs apps using the plugin `next-transpile-modules`.